Added configuration file for pgi compiler. Some minor fixes.
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@85 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/md_io/VISWriter.cpp
@@ -41,7 +41,7 @@ void VISWriter::doOutput( ParticleContainer* particleContainer,
       if(_incremental) {
          unsigned long temp = simstep/_writeFrequency;
 	 filenamestream << "-";
-         while(temp < floor(_numberOfTimesteps/_writeFrequency)){
+         while(temp < floor((double) _numberOfTimesteps / (double) _writeFrequency)){
             filenamestream << "0";
             temp = temp*10;
          }
